+++ algorithms/unit_tests/TestRandom.hpp
@@ -491,6 +491,34 @@ void test_random(unsigned int num_draws) {
 }
 }  // namespace Impl
 
+template<typename ExecutionSpace>
+void test_random_xorshift64() {
+#ifdef KOKKOS_ENABLE_SYCL
+        const int num_draws = 132141141;
+#else
+        const int num_draws = 10240000;
+#endif
+  Impl::test_random<Kokkos::Random_XorShift64_Pool<ExecutionSpace>>(
+      num_draws);
+  Impl::test_random<Kokkos::Random_XorShift64_Pool<Kokkos::Device<
+      ExecutionSpace, typename ExecutionSpace::memory_space>>>(
+      num_draws);
+}
+
+template<typename ExecutionSpace>
+void test_random_xorshift1024() {
+        #ifdef KOKKOS_ENABLE_SYCL
+        const int num_draws = 52428813;
+#else
+        const int num_draws = 10130144;
+#endif
+  Impl::test_random<
+      Kokkos::Random_XorShift1024_Pool<ExecutionSpace>>(num_draws);
+  Impl::test_random<Kokkos::Random_XorShift1024_Pool<Kokkos::Device<
+      ExecutionSpace, typename ExecutionSpace::memory_space>>>(
+      num_draws);
+}
+
 }  // namespace Test
 
 #endif  // KOKKOS_TEST_UNORDERED_MAP_HPP
